requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0 